Add drag & drop module
